requirements.txt
# Core dependencies
torch>=2.0.0
torch-geometric>=2.4.0
torch-geometric-temporal>=0.54.0
torch-scatter>=2.1.0
torch-sparse>=0.6.17
torch-cluster>=1.6.1
pyg-lib>=0.2.0

# Data and utilities
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
pyyaml>=6.0
omegaconf>=2.3.0

# Visualization and demos
matplotlib>=3.7.0
plotly>=5.15.0
streamlit>=1.25.0
pyvis>=0.3.2
networkx>=3.1

# Evaluation and logging
torchmetrics>=1.0.0
wandb>=0.15.0
tensorboard>=2.13.0

# Development tools
black>=23.0.0
ruff>=0.0.280
mypy>=1.5.0
pytest>=7.4.0

# Optional serving
fastapi>=0.100.0
uvicorn>=0.23.0
